refactor: move mocks into internal/
