+++ internal/mocks/module_repository.go
