+++ internal/server/v1/server_test.go
@@ -15,7 +15,7 @@ import (
 	"google.golang.org/protobuf/types/known/structpb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
-	"github.com/odpf/entropy/mocks"
+	"github.com/odpf/entropy/internal/mocks"
 	"github.com/odpf/entropy/module"
 	"github.com/odpf/entropy/resource"
 )
