+++ internal/store/inmemory/module_repository.go
