+++ resource/service_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 
-	"github.com/odpf/entropy/mocks"
+	"github.com/odpf/entropy/internal/mocks"
 	"github.com/odpf/entropy/resource"
 )
